working on tips screen
