+++ app/src/main/java/com/example/roundtimer/round_screen/presentation/TimerStatus.kt
@@ -1,4 +1,4 @@
-package com.example.roundtimer.round_screen
+package com.example.roundtimer.round_screen.presentation
 
 enum class TimerStatus {
     Running,
