--- app/src/main/java/com/example/roundtimer/setting_screens/TipsScreen.kt
@@ -1,24 +1,0 @@
-package com.example.roundtimer.setting_screens
-
-import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-@Composable
-fun TipsScreen(
-    onSwipeBack: () -> Unit
-) {
-    BackHandler {
-        onSwipeBack()
-    }
-
-    Column(modifier = Modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center) {
-
-    }
-}
