+++ app/src/main/java/com/example/roundtimer/setting_screens/tips_screen/TipsScreen.kt
@@ -1,0 +1,41 @@
+package com.example.roundtimer.setting_screens.tips_screen
+
+import androidx.activity.compose.BackHandler
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.systemBarsPadding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.roundtimer.common.BackIconButton
+
+@Composable
+fun TipsScreen(
+    onSwipeBack: () -> Unit
+) {
+    BackHandler {
+        onSwipeBack()
+    }
+
+    BackIconButton(
+        boxModifier = Modifier
+            .fillMaxWidth()
+            .systemBarsPadding()
+            .padding(16.dp),
+        boxAlignment = Alignment.TopStart,
+        onBackIconClick = onSwipeBack,
+        size = 40.dp,
+        iconTint = MaterialTheme.colorScheme.primary
+    )
+
+    Column(modifier = Modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center) {
+
+    }
+}
